Improved types and moved smart contract calls to services
