+++ src/services/smart-contract.service.ts
@@ -1,6 +1,7 @@
+import type { AppConfigCred } from '@/services/gateway.service'
 import signerMakeTx from '@/utils/signerMakeTx'
 
-async function setAppName(appName: string) {
+async function setAppName(appName?: string) {
   await signerMakeTx('setAppName', [appName])
 }
 
@@ -8,7 +9,7 @@ async function setDefaultLimit(storage: number, bandwidth: number) {
   await signerMakeTx('setDefaultLimit', [storage, bandwidth])
 }
 
-async function setClientIds(socialAuth: any[]) {
+async function setClientIds(socialAuth?: AppConfigCred[]) {
   const authSignerMakeTxValue: string[][] = [[], []]
   socialAuth?.forEach((authDetail) => {
     if (authDetail.verifier !== 'passwordless' && authDetail.clientId) {
